fix(문자표): symbols-dialog 테스트가 node 네이티브 러너에서 로드 가능하도록 순수 헬퍼 분리
symbols-dialog.test.ts 가 isCodePointInBlock 을 검증하려고 symbols-dialog.ts 를
직접 import 하면, 그 파일이 정적으로 가져오는 engine/command.ts 까지 전체가
node --test 의 TypeScript strip-only 모드로 파싱된다. command.ts 는 parameter
property 문법을 쓰는데, 이는 node의 strip-only 모드가 지원하지 않아
ERR_UNSUPPORTED_TYPESCRIPT_SYNTAX 로 죽는다.

isCodePointInBlock/UnicodeBlock 을 런타임 의존성이 없는 unicode-block.ts 로
분리하고, symbols-dialog.ts 는 그걸 재노출만 하도록 바꿔서 테스트가 command.ts
를 거치지 않고 순수 함수만 import 하게 했다. 이제 CI의 Frontend package gates/
Build & Test 가 통과한다.
